nas-server/requirements.txt
flask==3.1.0
pillow==11.2.1
